server/requirements.txt
Django==5.2
djangorestframework==3.16.1
django-cors-headers==4.8.0
mysqlclient==2.2.7

# Email
sendgrid==6.12.5

# JWT Authentication
djangorestframework-simplejwt==5.5.1

# Swagger/OpenAPI Docs
drf-yasg==1.21.10

# Misc dependencies
asgiref==3.8.1
Pillow==11.3.0
python-dotenv==1.0.1
pytz==2025.2
sqlparse==0.5.2
gunicorn==23.0.0
requests==2.32.5

# AWS S3 Storage
django-storages==1.14.2
boto3==1.35.84

# Formatting and Linting
black==25.1.0

# Debugging
django-debug-toolbar==6.0.0
django-extensions==4.1

# Testing/Fixtures
factory-boy==3.3.3
Faker==37.3.0
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-xdist==3.6.1
pytest-mock==3.14.0
pytest-env==1.1.5
pytest-randomly==3.16.0
responses==0.25.3
freezegun==1.5.1
coverage[toml]==7.6.9
model-bakery==1.19.5
